startup view with partially implemented functionality

+++ UI/ViewModels/Entities/MonitoredProcess.cs
@@ -22,7 +22,7 @@ public sealed partial class MonitoredProcess : ObservableObject
     => new()
     {
       Name = configuredProcess.Name,
-      AffinityValue = configuredProcess.AffinityValue,
+      AffinityValue = (nint)configuredProcess.AffinityValue,
       State = StateType.NotRunning
     };
 }
